container: Add a function to mount file systems
This is a preparation for the next patch where we move the logic to
moutn file systems from run_in_container to the container crate. This
patch just introduces a few functions in the container crate to mount
file systems, notably overlayfs. The code does almost the same as
run_in_container today. These functions will be called in the next
patch.

As a bonus, we get a few more unit tests.

BUG=b:280673549
TEST=bazel test //bazel/portage/common/container:all

Change-Id: I72b36bab006111dd457c42973dbda33b777dd4c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4719648
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>