container: Mount overlayfs before running run_in_container
Previously, run_in_container was responsible for mounting overlayfs and
bind-mounts. With this change, the container crate is responsible for
them. Now run_in_container just pivot-roots into the specified directory
after setting up a sandbox with various namespaces.

This change allows the users of the container crate to inspect the
container file system directly. The main motivation of this change is to
allow post-processing the overlayfs upper directory to clean up
unnecessary files and reducing the number of file system layers.

BUG=b:280673549
TEST=bazel test //bazel/portage/common/container:all
TEST=BOARD=amd64-generic bazel build @portage//sys-apps/acl

Change-Id: Ib39ccf27e1beb865c340dba65fb08103d169e400
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4719649
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>