Fix ordering of eclass paths in Alchemy
Eclass search paths should be the inverse of the overlay's "masters"
order, to respect the intended semantics:

  "The masters list is in reverse priority order: later entries take
  precedence."

This commit fixes the divergence of the Alchemy's eclass inheritance
simulation from the real Portage's one. Note that it only makes
difference for same-named eclasses located in different places, which in
practice only means the fix is relevant for "toolchain-funcs.eclass"
(for the correct selection between its portage-stable and
chromiumos-overlay implementations).

BUG=b:293522502
TEST=add debug logging about inherited eclasses paths, build chromeos-base/chromeos-config-bsp, verify "chromiumos-overlay/eclass/toolchain-funcs.eclass" is used (and not "portage-stable/eclass/toolchain-funcs.eclass")

Change-Id: I6b33030fc1c3d66d592f51bff4c4155ecf790f47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4729041
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>