durabletree: Introduce "hot" state
This patch fixes the problem where durable trees fail to restore file
permissions by introducing the "hot" state to durable trees. It also
adds regression tests.

What is the hot state?

When a program creates a durable tree with [`DurableTree::convert`], it
is initially marked as *hot*. A durable tree is considered hot if its
root directory has a permission different from 0555.

It is an error to attempt to expand a hot durable tree with
[`DurableTree::expand`] because it would mark the tree "already
restored", while Bazel would modify permissions of files in the tree
after the current Bazel action finishes, which results in corrupted
permissions.

Therefore, a Bazel action should convert its output directory to a
durable tree at the end of its execution and should never attempt to
expand it. Once the action finishes, Bazel modifies permissions of the
durable tree to 0555, which essentially removes the hot state. From
this point, Bazel won't modify file permissions anymore, so it is safe
to expand the durable tree and set the "already restored" marker in
other actions.

If you need to expand a hot durable tree in unit tests, you can use
[`DurableTree::cool_down_test_only`] to simulate the "cool down" process
of Bazel.

BUG=b:294159699
TEST=bazel test //bazel/portage/common/...

Change-Id: Ie465dc19d123530a43afa59f11d1d422cab039fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4741017
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>