Pass metadata to the real elf interpreter.
This can, for example, be read by the elf interpreter to ensure that we attempt to load shared libraries from $SYSROOT/lib/libc.so instead of /libc/libc.so.

This is preferred over just adding the entries to rpath, since this
method ensures that you never fallback to the one installed on the host
system.

BUG=b:291840720
TEST=portage/tools/run_tests.sh

Change-Id: I4d1c1aa5a0def0e8d78fe8491c1a6d714246e3fc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4736975
Tested-by: Matt Stark <msta@google.com>
Commit-Queue: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>