alchemist: Add workaround for amd64-host PORTDIR_OVERLAY ordering
The RepositorySet impl makes an assumption that the "primary" (overlay
that contains the board's profile) is always the last one listed. This
convention holds true for public builds and internal `target/board`
builds. It doesn't hold true for internal amd64-host builds though. This
is because the amd64-host /etc/make.host_setup file appends to the
PORTDIR_OVERLAY resulting in the `chromeos` repository being the last
one.

This CL adds a workaround that filters out the "chromeos" and
"chromeos-partner" repositories when computing the primary repository.
This results in getting the expected `amd64-host` repository. I plan on
doing some refactoring to fix this correctly, but this will unblock
building inside the chroot.

BUG=b:293383461
TEST=(chroot) BOARD=arm64-generic bazel build @portage//internal/packages/stage2/host/chromiumos/sys-libs/libselinux:3.0-r5

Change-Id: Ia8f23a5c3112419c0090ac906ec51f42c65f2b18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4753639
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Ryo Hashimoto <hashimoto@chromium.org>