portage: Mask 9999 llvm-project ebuilds
We don't want to build the 9999 variants of the llvm-project ebuilds
because they require the whole .git directory. This causes problems
because the .git repository isn't a stable input and results in cache
busting everything on every sync. If these packages weren't primordial
it wouldn't be that big of a deal. We plan to address building 9999
llvm-project ebuilds later down the road.

The only 9999 llvm-project ebuild that currently exists is libcxx. The
others are still being migrated. This change adds them all just so we
don't encounter any problems later.

BUG=b:267209587, b:280673549
TEST=BOARD=arm64-generic bazel build -k @portage//internal/sdk/stage2/target/board:aarch64-cros-linux-gnu

Change-Id: I4e522be1b3bd90c675b3cf2c3d16615d952a3922
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4798915
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Tested-by: Raul Rangel <rrangel@chromium.org>