alchemist: Generate source packages relative to repo root
This change drops the need for the special @chromite repository, and
instead utilizes the existing source BUILD generation logic we have
in place. This required re-rooting the //internal/sources so that
it contains paths relative to the repo root, instead of src/.

We can probably clean up the API for generate_repo_main to take in a
repo_root instead of a src_dir, but that would have made this CL even
messier. We can do that as a followup.

BUG=b:297534010
TEST=BOARD=arm64-generic bazel build @portage//internal/packages/stage1/target/board/chromiumos/virtual/target-os:1_package_set
TEST=BOARD=arm64-generic bazel query @portage//internal/packages/stage1/target/board/chromiumos/chromeos-base/hwid_extractor/...

Change-Id: I861c24cd4a741a19ec16a51ddb171cbfdae49c50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4814583
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>