Generate a manifest of a tbz2 file to be checked in.
tbz2 files are extracted during the execution phase. Thus, it is
impossible to know the contents of a tbz2 file before the execution
phase.

To solve this issue, we have a constant list of files we expect to see
at the analysis phase. During the execution phase, we verify that it
matches what we actually got.

Thus, if the tbz2 file ever changes the set of files it produces (not
the contents, only the paths), we will fail to compile and tell the user
a command to run that fixes the problem.

BUG=b:265990534
TEST=portage/tools/run_tests.sh

Change-Id: I40a3b7442624add7f47e705d5acadbd885ddbfc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4835493
Auto-Submit: Matt Stark <msta@google.com>
Tested-by: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Matt Stark <msta@google.com>