portage: Update to cros-sdk-2023.09.08.050046
python3.6 is misbehaving because it's looking for OpenSSL 3 in the BROOT
instead of the ROOT when building the stage2 SDK. Revbumping the SDK
works around this problem because it pulls in OpenSSL 3 into the SDK.
Since we are dropping python 3.6 soon, this workaround is the least
amount of work. Python 3.8 doesn't suffer this problem.

BUG=b:299620346
TEST=BOARD=amd64-generic bazel build -k @portage//internal/packages/stage2/target/board/portage-stable/dev-libs/openssl:3.1.2
TEST=BOARD=amd64-generic bazel build -k @portage//dev-libs/openssl:3.1.2

Change-Id: I882ff299ed721c3c46795b471e058f0b4806199c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4854484
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Tim Bain <tbain@google.com>
Reviewed-by: Tim Bain <tbain@google.com>
Auto-Submit: Raul Rangel <rrangel@chromium.org>