Improve support for regular files in CROS_WORKON_SUBTREE.
Instead of depending on a full contents of the directory where the file
is located, we generate a new target, __single_files_tarball_,
which contains directly referenced files in that directory.

This patch will remove the dependency on the full contents of
platform2 when platform2/.gn is required.

BUG=b:272407670
FIXES=b:272407670
TEST=BOARD=amd64-generic bazel build --@portage//:default-stage=stage1 @portage//chromeos-base/hammerd
TEST=BOARD=amd64-generic bazel build --@portage//:default-stage=stage1 @portage//virtual/target-os:package_set

Change-Id: Ic579819ea6deb71d6ba69130ebbd2f0dd7e9b0db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4867752
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>