Investigate which process calls chmod on hot durable trees
trapfs is a simple FUSE file system that mounts an empty file system and
waits up to 10 seconds until some process call chmod on the directory.
When chmod is called, info about the process is printed to stdout.
trapfs is written in Go because I could not find a good Rust FUSE
library that is capable of inspecting caller process IDs. In any case,
we won't keep trapfs longer than O(days).

We modify action_wrapper to scan durable trees specified in input layer
arguments and perform hotness checks. If it finds hot durable trees, it
installs trapfs on its subdirectory and waits for a bad process to call
chmod on this.

Since we need a directory under durable trees for mount points, this
patch also adds one named "trapfs_mount_point_b299934607". This is only
for temporary use and will be removed together with other hacks once we
finish debugging the issue.

BUG=b:299934607
TEST=bazel build //bazel/portage/bin/trapfs
TEST=BOARD=amd64-generic bazel build @portage//sys-apps/attr

Change-Id: Ic29a58c0ed864438735d357cade5f4d3077cd91f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4876498
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>