trapfs: Workaround deadlock
For some reasons, server.Wait causes deadlock. I'm suspecting that this
is because the FUSE mountpoint is still alive in other mount namespaces,
but I'm not 100% sure. In any case, after 10-second timeout, we don't
have reasons to serve the file system, so let us forcibly exit the FUSE
process. This may put left mountpoints inaccessible, but we don't care
about them as the build is doomed to failure.

BUG=b:299934607
TEST=bazel run //bazel/portage/bin/trapfs mountpoint

Change-Id: If6e1c110beee1c7c72fa9a6c32d2ba75008cd391
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4881677
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>