package_contents: Accept empty strings to generate contents for host
This is a reattempt of "Fix host package builds with fast package
installation" (crrev.com/c/4875900).

I thought the default value for unspecified string attributes is None,
but it was an empty string instead, which caused errors on analysis
time.

An earlier failed attempt made the board attribute for ebuild* rules
mandatory, and defined to accept "amd64-host" as the board name on
building for host. But this introduced undesired side effects because
we can no longer distinguish installing to the host (/) and installing
to the amd64-host sysroot (/build/amd64-host).

This patch leaves the definition of ebuild* rules as-is, but instead
fixes generate_contents() to accept empty board names, which means to
install to the host. Also fixes types in relevant code.

BUG=b:300130101
TEST=BOARD=amd64-generic bazel build \
     --@portage//:binpkg-cache-strategy=fast @portage//sys-apps/attr
TEST=BOARD=amd64-generic bazel build \
     --@portage//:binpkg-cache-strategy=naive @portage//sys-apps/attr

Change-Id: I59b1e8d058d52689e241bc38fbbdb576679199c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4896616
Auto-Submit: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>