Statically determine a sysroot path to install a package at
Previously, fast package installation was designed to support installing
host packages to both / and /build/amd64-host. However this is not
necessary as of today: when we build host packages in stage1, they will
be installed to /build/amd64-host, and when we build host packages in
stage2, they will be installed to /.

This patch reduces Starlark complexity by introducing an assumption that
a binary package built with Bazel will be installed to a single sysroot
path.

BUG=b:300130101
TEST=BOARD=amd64-generic bazel build \
     --@portage//:binpkg-cache-strategy=fast @portage//sys-apps/attr
TEST=BOARD=amd64-generic bazel build \
     --@portage//:binpkg-cache-strategy=naive @portage//sys-apps/attr

Change-Id: I99ef81b6e0db5c5a6ee5a910c08c7ee8a727aaba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4899557
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>