Rename @portage//:* to //bazel/portage:*
Placing a configuration in a repo means that if you want it in your
bazelrc, then the repo always needs to be generated, even if it's not in
use. This would mean that, for example, in order to run
//bazel/python:python_test, I would need to generate @portage
unneccesarily.

BUG=None
TEST=portage/tools/run_tests.sh

Change-Id: Ic39e0272df666cd19e7592462cb640ac96f4b364
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4904662
Auto-Submit: Matt Stark <msta@google.com>
Commit-Queue: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Matt Stark <msta@google.com>