portage: Don't fail if patches don't apply
In order to land the patches in our portage fork, we need to make sure
we don't fail if the patch has already been applied. Unfortunately we
don't have a reliable way to determine if the patch has been already
applied, or if there was a merge conflict :/ I guess it's more
motivation to push all the patches up.

BUG=none
TEST=BOARD=amd64-generic bazel build @portage//internal/sdk/stage2

Change-Id: Ia94a65a2ae5d8d4bc06a40c810486841512d94cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4907379
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>