Remove duplicate Cargo.toml
Now that we aren't sharing the Cargo.toml with the the entire workspace,
we can drop the duplicate alchemy_crates Cargo.toml. The Cargo.toml were
a bit out of sync. I tried to use the root Cargo.toml as the source of
truth, but `cargo-bazel` doesn't support the `default_features`
attribute. So I dropped it since the alchemy_crates didn't have it
specified.

This change does mean we need to keep the workspae members in sync, but
cargo-bazel will raise an error if they aren't, so its a lot easier
to keep in sync then the two Cargo.tomls.

We might also be able to move alchemist cargo workspace back into the
the bazel cargo workspace now.

We should probably also evaluate migrating to third_party/rust_crates
and using that workflow so we can get cargo-vet support.

I needed to add a Cargo.toml to the examples directory because
its' the only crate using `env_logger` and `log`. With out the
Cargo.toml we won't generate these crates as repos so the example
bazel build will fail.

BUG=none
TEST=BOARD=amd64-generic bazel build --config=prebuilts/stage2-board-sdk @portage//target/chromeos-base/libchromeos-use-flags
TEST=~/chromiumos/src/bazel/portage/bin/xpaktool$ cargo test
TEST=./bazel/portage/tools/run_tests.sh

Change-Id: I96f4824cfdd5e7ac52408f2c4a325b72905fdd24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4919604
Reviewed-by: Matt Stark <msta@google.com>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>