gs: Execute out of workspace root
`gsutil` expects to be executed inside the CrOS src checkout. This is
because it uses it to locate the cache directory. When it's outside of
this directory it will use `/tmp`, and that will sometimes get wiped
leaving the cache in an invalid state.

BUG=b:303375562, b:303890102
TEST=./bazel/portage/tools/run_tests.sh

Change-Id: I9faca1974c890fd91e38509cb6b110706143624b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4919768
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Tim Bain <tbain@google.com>