alchemist: Make FEATURES partially incremental
Portage lists FEATURES as an incremental variable, but it has a
non-blank default. This means we can't fully compute the FEATURES flag
in alchemist because we don't know what the defaults are. Instead we
will treat the FEATURES variable as a partially incremental variable.
i.e., we won't resolve any of the tokens, but instead just concat
all the values so that portage can compute the final value.

The `news` feature was being enabled when I started injecting the
compiled make.default.

BUG=b:262460688
TEST=BOARD=amd64-generic bazel test @portage//internal/packages/stage1/target/host/chromiumos/sys-kernel/linux-headers:4.14-r91_portage_profile_test
TEST=cargo test

Change-Id: Iaa20a9812cba7b5932bb8dca76a79d917d41d5e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4924200
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>