Replace a list of binpkgs with a list of BinaryPackageInfo.
In a follow-up CL, update_manifest will also be able to generate a
remote manifest which is used to generate a repo rule.

The repo rule will need to create a build file along the lines of:
binary_package(
    name = "binpkg",
    package_name = {package_name},
    src = {src},
    category = {category},
    slot = {full_slot},
    version = {version},
    visibility = ["//:__subpackages__"],
    runtime_deps = {runtime_deps},
)

This requires knowing all of the info in BinaryPackageInfo, rather than
just the name of the file.

BUG=b:265990534
TEST=portage/tols/run_tests.sh

Change-Id: Ifca6943f023a1806cab435f6a28aa4346913efbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4923230
Tested-by: Matt Stark <msta@google.com>
Auto-Submit: Matt Stark <msta@google.com>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Matt Stark <msta@google.com>
Tested-by: Tim Bain <tbain@google.com>