Add the ability to create a tarball including a binary and all its runfiles.
This is required to break a dependency loop.

The extraction package relies on having a toolchain configured (ie. primordial toolchain), because it's a rust binary.
To configure the primordial toolchain, we need to run the extraction
binary to extract the contents of the toolchain from the tbz2 file.

BUG=b:265990534
TEST=portage/tools/run_tests.sh

Change-Id: Ifacb77858136012224525de87362eb6d52398307
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4930969
Tested-by: Matt Stark <msta@google.com>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Matt Stark <msta@google.com>