images: Don't abort when BUILD.bazel is loaded without $BOARD
Currently Bazel fails to load //bazel/images:BUILD.bazel when $BOARD is
not set, even if we don't build any targets there, because it results in
passing None to the mandatory attribute of the build_image rule. This is
inconvenient when we use wildcards to capture targets, e.g.
`bazel test --build_tests_only //bazel/...`.

This patch introduces a helper function that conditionally calls a rule
function when $BOARD is supplied, which fixes the problem.

BUG=b:304630955
TEST=bazel test --build_tests_only //bazel/...
TEST=bazel test //bazel/images:chromiumos_base_image
TEST=BOARD=amd64-generic bazel test //bazel/images:chromiumos_base_image

Change-Id: I2e03eb5ceaabebd3368afab939e6156c05058e99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4937931
Reviewed-by: Tomasz Tylenda <ttylenda@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>