alchemist: Use a dedicated target directory when building alchemist
The wrapper was previously sharing the target directory with the
regular `cargo` workflow. i.e., `cargo build` or an IDE. This means
the binary can get ripped out from under bazel while trying to execute
alchemist. This change makes it so we use a dedicated .cache directory
that only bazel has access to.

I also took the opportunity to refactor the script a bit. Previously
it was also symlinking the `CACHEDIR.TAG` file, but this should never
be a symlink according to the spec. https://bford.info/cachedir/
The new method only symlinks the `<profile>` directory.

I also got rid of the `set +e` since we want to make sure to fail the
script if creating the symlink fails.

BUG=b:303391385
TEST=bazel build //bazel/portage/sdk:extra_tarball
TEST=tree /home/rrangel/.cache/bazel/_bazel_rrangel/8bde7ab3187096959038758eb972b27a/external/_main~portage~alchemist/

Change-Id: I3150fb2fd8ccbc9ed6b69960797749b967f55729
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4936958
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Tim Bain <tbain@google.com>
Auto-Submit: Raul Rangel <rrangel@chromium.org>