alchemist: Introduce EBuildBasicData
EBuildBasicData contains basic information about an ebuild, that is
available as long as an ebuild file exists with a correct file name
format. This struct will be used in later patches to provide the fields
common to all package-related structs, such as ebuild::PackageDetails
and generate_repo::common::Package.

This is a pure refactoring and not supposed to change any behavior.

BUG=b:303400631
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: Ief467bcfff2e6076306dac323f5d2ea49bc1c537
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4982813
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>