alchemist: Introduce MaybeEBuildMetadata
Formerly, EBuildMetadata represented both successful/failed cases of
ebuild evaluation. It was good in the sense that we can handle both
cases gracefully, but bad in the sense that we cannot statically
guarantee that an EBuildMetadata corresponds to a successful case.

This patch introduces MaybeEBuildMetadata, a simple enum pointing to
EBuildMetadata or EBuildEvaluationError, which solves the problem.

This is a pure refactoring and not supposed to change any behavior.

BUG=b:303400631
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I7b67c554e60bae46e6f9fd49f8e3d87f73cb3ef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4982814
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>