[alchemist] Refactor metadata and analysis errors
Refactor the PackageError to split it into different types
depending on the particular type of the error.

This allows for a cleaner data model for different error scenarios
and avoids confusion: e.g., if a failure happened during the ebuild
parsing we have substantially less bits of information than for
failures happening during the dependency analysis stage.

BUG=b:303400631
TEST=SKIP_PORTAGE_TESTS=1 portage/tools/run_tests.sh

Change-Id: I9a7150fefc9c1a8f25231c4ca9d0c5d33df2140d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4957013
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>