alchemist: Break DependencyParserCommon into utility functions
Formerly we had the DependencyParserCommon trait that implemented common
logic for parsing dependency expressions. But this method was tricky as
the responsibilityf of trait was unclear except that it had "common"
logic. Also, now that we're adding a new type of composite dependencies
soon, this design makes it difficult to reuse the logic.

This patch breaks the common logic in DependencyParserCommon into helper
functions. Instead it introduces the PartialExpressionParser trait that
abstracts a partial expression parsing function. We can call helper
functions by providing a type implementing the trait.

BUG=b:268258415
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I078d14abaf0a68aedbc92ac49f485dc11f0e09c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4958538
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>