alchemist: Make UseMap a new type
Previously UseMap was an alias to HashMap, but it turned out that it's
inconvenient that it doesn't implement AsRef<UseMap>. According to the
standard library doc, AsRef is not reflexive for technical reasons:
https://doc.rust-lang.org/std/convert/trait.AsRef.html#reflexivity
The doc sugggests to implement AsRef<T> for T where necessary, but we
can't do it when UseMap is an alias to HashMap.

This patch makes UseMap a new type to implement AsRef<UseMap>. We though
try to make it behave like a plain HashMap by implementing Deref and
DerefMut.

BUG=b:268258415
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I3caa090ef4d76321dc600f0beac3396ab65c0fa9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4958540
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>