alchemist: Introduce MaybePackageDetails
This patch follows the same pattern as MaybeEBuildMetadata: introduces
an enum that allows interleaving success/failure cases of package
loading and differentiating them statically at the same time.

While this new enum allows us cleaning up a few logic, this patch
focuses on introducing it with minimal changes.

This is a pure refactoring and not supposed to change any behavior.

BUG=b:303400631
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: Ie0d1396c7a068b697b3e404aa8dcd287465a0b3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4982815
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>