alchemist: Introduce portage_config generator
In order to support building packages using compiled profiles and
"full" portage profiles, we need to extract the portage config from the
SDKs and instead inject it at the package level. This change introduces
a new @portage//internal/portage-config directory that will house all
the portage configs for all the various build targets. I'm only starting
with the `host` target for now to lay the infrastructure down.

The `host:orig` profile is currently hard coded because it's the profile
that is currently built into the sdk tarballs. That means it's the
default profile the host uses when doing a cross-root build. This was
always one very subtle thing that is easy to miss when I implemented the
host tool builds. I plan on migrating all `target` builds to using a
`host:lite` profile instead so we can drop the dependency on the
`make.conf.amd64-host` and `default/linux/amd64/10.0/sdk` profiles. The
lite profile will only have `CHOST` and `ARCH` defined since that's
all that's required when doing a cross-root build.

The `host:full` profile is like `orig`, but it uses the `bootstrap`
profile. This is the profile we use when building `ROOT=/` packages.

BUG=b:262460688
TEST=BOARD=amd64-generic bazel build @portage//host/dev-util/cmake @portage//host/dev-vcs/git

Change-Id: Idf5b7ebb1cee4e6b359f9222c7195e9e1e085be1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4998850
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>