alchemist: Fail on matching a block dependency atom with a single package
We can't define a reasonable behavior on matching a block dependency
atom with a single package. This patch makes it an error to perform such
checks.

Ideally, we should introduce a new type that is similar to
PackageDependencyAtom but guaranteed to contain no block. Then we can
move these matcher methods to the new type and drop the error handling.
I actually tried it a bit, but it turned out to be a large patch despite
its limited usefulness, so I'm just leaving a TODO.

BUG=b:299554821
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: Ic3d9737664fe2f22b381d69ce16409448ff07015
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/4996805
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>