alchemist: Add base profile to test data
We will use this profile to verify that the profile compiler generates
the expected outputs. It also adds test coverage for the USE_EXPAND
logic. The base profile was copied and stripped down to not cause an
excessive amount of USE flags to be generated.

BUG=b:262460688
TEST=cargo test

Change-Id: Ied8ec8d43c90742f03540439a3fae6b0234b0b8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5011321
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>