alchemist: Generate compiled profiles for targets
This change introduces a "config compiler" that takes a config bundle
and generates a portage config. The compiler doesn't do anything fancy.
It just splits the config bundle into make.conf and make.defaults.

I'm not generating a profile for the host yet because the config bundle
we have is for the cross-root amd64-host, so it contains a bunch of
cross-root variables (e.g., ROOT=/build/amd64-host). We need to strip
these out and rewrite some of them. I'm punting on this for now.

BUG=b:262460688
TEST=cargo test

Change-Id: Ia7d945b1ab79ab37fae7124f31d4f82cc3896f6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5011322
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>