alchemist: Resolve ../ in profile paths
This change introduces a new `clean_path` method that will resolve the
../ in a path. The reason this is necessary is because we need to start
passing the absolute path to the profile.bashrc and a package's bashrcs
into the build_pacakge container. Without this change we pass in paths
like:
    /mnt/host/source/src/third_party/chromiumos-overlay/profiles/default/linux/amd64/10.0/sdk/../../../../../base/profile.bashrc

This ends up being problematic once the `parent` files are removed from
the overlays because the intermediate profile directories will no longer
exist in the container. With this change we generate the following path:

    /mnt/host/source/src/third_party/chromiumos-overlay/profiles/base/profile.bashrc

Why not use `path.canonicalize()`? When running outside the chroot, the
fakechroot module will create a fake root and pivot to it. This means
that `canonicalize` will return the `/.old-root/` path. It will also
include the $USER in the path. We're not really interested in the real
path, but just a clean one we can use inside the container.

BUG=b:262460688
TEST=cargo test

Change-Id: I76776737f57203cc96a10ec9d1823303da909db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5027784
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>