alchemist: Teach alchemist about bashrc files
My goal is make it so that  packages compiled in different
configurations (i.e., amd64-generic vs amd64 generic private) might
de-dupe to the same action keys. Dropping the make.profile symlink is
a part of that work. Now that make.defaults and package.use are part of
the compiled profile, the only significant thing left that we get from
the profiles is the bashrc configuration.

This change teaches alchemist how to parse the profile.bashrc and
package.bashrc and adds the matching bashrc's onto the Package. The list
stored in Package is the order that the bashrc files need to be
executed. Because each profile node can define profile.bashrc and
package.bashrc, they could be intermixed.

Rendering the paths will be done in a follow up.

BUG=b:262460688
TEST=cargo test

Change-Id: I912611fe0f46edd0b179f08c5225338c91784f7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5027785
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>