portage: Add a bashrcs argument to ebuild target
This change allows specifying the list of bashrc (profile.bashrc and
the package's bashrc) that need to be executed when building the
package. The reason it's package specific is because `package.bashrc`
allows specifying multiple package atoms, so that means we need a
per-package bashrc list.

This API is subject to change. Right now it assumes that all the bashrc
files and their dependencies are included in the layers. I might evolve
this API to bind mount the bashrc and remove it from the layers, but
that leaves the question of how to specify resources. Right now none
of the `profile.bashrc` or `package.bashrc` files have any resources,
but I want to plan ahead for it because ideally we can remove the custom
CrOS `bashrc` logic from the `base/profile.bashrc`, and instead use the
portage `package.env` mechanism. This would allow us to apply cross-repo
`bashrc` files and clean up the chromiumos overlay tarball.

BUG=b:262460688
TEST=BOARD=amd64-generic bazel build --//bazel/portage:compiled-profiles @portage//target/app-crypt/rhash

Change-Id: I6db41701c52d9285fcd7c74107227bf1712fb7f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5027786
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>