alchemist: Disable find-missing-deps.sh from bazel builds
This script doesn't work inside the bazel container because it fails to
find libc.so.6. This happens because we never install the glibc package
into the target/board sysroot, but instead manually unpack the binpkg
and copy the contents into the sysroot. The find-missing-deps.sh script
has a work around for this where it checks to see if the host has the
cross-XXX/glibc package installed instead. We don't install the
cross-XXX/glibc package in the board's toolchain layer because it's not
necessary to build any packages.

Since we are working on removing chromite dependencies, I think it's
better to just exclude the find-missing-deps.sh hook.

I had to hack the source generator to do this. My first attempt was
adding an exclude field to the PackageLocalSource::Src enum, but that
makes the API very confusing because of the nested nature of the
generated source files.

BUG=b:301462588
TEST=BOARD=amd64-generic bazel build @portage//internal/sources/src/scripts/hooks:__tarballs__

Change-Id: If43df4107d13e9df5b4d306aa0ca35d014cfc8c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5034773
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Raul Rangel <rrangel@chromium.org>