alchemist: Move compiled profile make.conf to /etc
It turns out that the license_lib uses the sysroot lib to query the
PORTDIR_OVERLAY variable. The sysroot lib expects the make.conf file to
exist at /etc/make.conf. Ideally licensing could just invoke `portageq`
to extract this information. The problem was only showing up for some
target packages because most licenses are stored in the portage-stable
overlay. The licenses that live in chromiumos-overlay were the packages
that were showing this problem.

For now let's move the file into the legacy location.

BUG=b:262460688
TEST=BOARD=amd64-generic bazel run --//bazel/portage:compiled-profiles @portage//target/chromeos-base/chromeos-ca-certificates:debug

Change-Id: I624ae1784a628c8f56b8cbbbd3cb5e97776ba110
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5034772
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>